Adds work-around for ruby issues.
I suspect that the ruby development libraries (listed as 1.9.1) are out
of date with the current ruby scripts (1.9.3).  I added a dummy
synchronize method to the Mutex class in order to get around the errors
I was seeing when instantiating ruby from C, and I'm crossing my fingers
that nothing breaks.

I also changed my development environment from the default Ubuntu
packages to RVM and updated the Makefile to point to the new headers and
shared libraries.